Wow. Why didn't I commit this in smaller parts?
Lots of changes. This started as a small change to split the classes in
ib3 into a few sub-modules. That led to fixing up the generated
documentation a bit. Then I just kept running in a circle of
refactoring, updating docs, and creating new sub-modules.

The neat thing though is that the library is looking pretty useful and
the example saslbot.py class does things that none of the live bots
I started extracting this from can do yet.